requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
sqlparse==0.5.0
tzdata==2024.1
